Scraping complete. More logic in main. Fixed ruff!
Completed the scraping process. This needs to be looked over again with
fresh eyes. But for now it works fine.
Implemented more logic in main. I have the process in my head but it's
time for lunch.
Started working on custom exceptions. Might have to read up some more
before continuing. But you know.. Works.